requirements.txt
streamlit
torch
torchvision
matplotlib
numpy
opencv-python
seaborn
Pillow
scikit-learn
